@AndyAyersMS 2018-12-13
Keeping this as [WIP] for now since I'd like to see some of the other R2R and tiering perf changes land first, and evaluation of this change may be tricky, and nothing depends on this. And I'll be on vacation a fair amount in the coming weeks.

I don't expect 16 byte alignment to universally improve perf as whatever method alignment we get now on a given run may work out better. But over time this alignment should be more perf stable and also gives the jit the opportunity to internally pad to avoid bad instruction fetch issues for hot loops and eventually end up in an overall better place.

I plan to make multiple measurements on different machines and over time to try and show results are indeed more stable and (for benchmarks that consistently stand out as better or worse) drill in and try to understand why.

I don't think this will cause much extra code space fragmentation -- though aligning small methods may have a notable impact. So another goal is to actually measure the size impact, especially on realistic apps.

Using the performance repo I have done a couple of base and diff runs of local builds on a Skylake  i6700 and compared them . Data below is filtered by absolute difference of at least 1ms as results for shorter running tests seem to swing wildly. @adamsitnik if you want to experiment with this change and see why there's such volatility in the shorter tests I'd love to see any insights you uncover.

cc @dotnet/jit-contrib @fiigii 

## BASE vs DIFF 1

| Slower                                  | diff/base | Base Median (ns) | Diff Median (ns) | Modality              |
| --------------------------------------- | ---------:| ----------------:| ----------------:| ---------------------- |
| LinqBenchmarks.Order00ManualX           |      1.13 |     128679250.00 |     145741725.00 |                       |
| SciMark2.kernel.benchMonteCarlo         |      1.06 |     720936900.00 |     761832900.00 | can have several modes|
| LinqBenchmarks.Where00ForX              |      1.04 |     407904300.00 |     424595200.00 |                       |
| BenchmarksGame.BinaryTrees_2.RunBench   |      1.04 |     100067050.00 |     103815100.00 |                       |
| ByteMark.BenchNeural                    |      1.03 |     684599100.00 |     706311100.00 |                       |
| CscBench.DatflowTest                    |      1.03 |     375537650.00 |     386855100.00 |                       |
| Benchstone.BenchF.NewtE.Test            |      1.02 |     396415000.00 |     405177150.00 | bimodal               |
| Benchstone.BenchF.Regula.Test           |      1.02 |     238591750.00 |     243644100.00 |                       |
| LinqBenchmarks.Count00ForX              |      1.02 |     328557500.00 |     334972000.00 |                       |
| LinqBenchmarks.Where01LinqMethodNestedX |      1.01 |     367130500.00 |     370483800.00 |                       |
| LinqBenchmarks.Count00LinqMethodX       |      1.00 |     829226900.00 |     832690600.00 |                       |

| Faster                          | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| ------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| Burgers.Test3                   |      2.54 |    1063852600.00 |     418235300.00 |         |
| SciMark2.kernel.benchFFT        |      1.11 |     767675800.00 |     694001500.00 |         |
| SeekUnroll.Test(boxedIndex: 27) |      1.02 |    1679506400.00 |    1649932800.00 |         |
| Benchstone.BenchF.LLoops.Test   |      1.02 |     566127600.00 |     556443250.00 |         |
| Benchstone.BenchF.FFT.Test      |      1.02 |     172081000.00 |     169311400.00 |         |
| SeekUnroll.Test(boxedIndex: 19) |      1.02 |    1358964600.00 |    1338244400.00 |         |

## BASE vs DIFF 2

 Slower                                                   | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| Benchstone.BenchF.DMath.Test                             |      1.08 |     651020750.00 |     701478200.00 |         |
| SIMD.ConsoleMandel.VectorFloatSinglethreadADTNoInt       |      1.06 |     276949100.00 |     293837350.00 |         |
| LinqBenchmarks.Order00LinqQueryX                         |      1.05 |      88074675.00 |      92741887.50 |         |
| LinqBenchmarks.Count00ForX                               |      1.05 |     328557500.00 |     344875100.00 |         |
| PerfLabTests.DelegatePerf.MulticastDelegateCombineInvoke |      1.03 |     206379175.00 |     212145600.00 |         |
| LinqBenchmarks.Where00ForX                               |      1.03 |     407904300.00 |     419107750.00 |         |
| ByteMark.BenchNumericSortJagged                          |      1.02 |    1182589800.00 |    1209204000.00 |         |
| Benchstone.BenchF.Regula.Test                            |      1.02 |     238591750.00 |     243943800.00 |         |
| ByteMark.BenchFourier                                    |      1.02 |     483880550.00 |     492823600.00 |         |
| Benchstone.BenchF.NewtE.Test                             |      1.01 |     396415000.00 |     402287800.00 |         |
| SciMark2.kernel.benchMonteCarlo                          |      1.01 |     720936900.00 |     731035700.00 |         |
| LinqBenchmarks.Where00LinqQueryX                         |      1.01 |     534897850.00 |     541721300.00 |         |
| CscBench.DatflowTest                                     |      1.01 |     375537650.00 |     379996200.00 |         |
| ByteMark.BenchEmFloat                                    |      1.01 |    3250239200.00 |    3273381850.00 |         |

| Faster                                                                                               | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| ---------------------------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| Burgers.Test3                                                                                        |      2.53 |    1063852600.00 |     420375300.00 |         |
| SciMark2.kernel.benchFFT                                                                             |      1.16 |     767675800.00 |     663166950.00 |         |
| Benchstone.BenchI.BubbleSort2.Test                                                                   |      1.05 |      30900981.25 |      29298325.00 |         |
| Benchstone.BenchI.NDhrystone.Test                                                                    |      1.03 |     356696700.00 |     347896800.00 |         |
| BenchmarksGame.MandelBrot_7.Bench(size: 4000, lineLength: 500, checksum: "C7-E6-66-43-66-73-F8-A8-D3 |      1.02 |     117919350.00 |     115333875.00 |         |
| Benchstone.BenchF.LLoops.Test                                                                        |      1.02 |     566127600.00 |     556375000.00 |         |
| LinqBenchmarks.Where01LinqQueryX                                                                     |      1.01 |     320124900.00 |     316781200.00 |         |
| SIMD.RayTracerBench.Bench                                                                            |      1.01 |     470632650.00 |     467229150.00 |         |
| SeekUnroll.Test(boxedIndex: 27)                                                                      |      1.01 |    1679506400.00 |    1669661100.00 |         |
| SeekUnroll.Test(boxedIndex: 19)                                                                      |      1.01 |    1358964600.00 |    1351673500.00 |         |
| Benchstone.BenchF.Whetsto.Test                                                                       |      1.00 |     529907750.00 |     527334600.00 |         |